% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\name{atom}
\alias{atom}
\title{Build an atom
From the number of protons and neutrons, report the corresponding element/isotope, its abundance, half-life, decay mode and daughter product.}
\usage{
atom(protons = c(), neutrons = c(), roundby = 1, talk = TRUE)
}
\arguments{
\item{protons}{The number of protons in the atom's nucleus. Alternatively, the name or symbol of an element can provided.}

\item{neutrons}{The number of neutrons in the atom's nucleus. If left empty, the function will report the amounts of neutrons of the isotopes of the element in question.}

\item{roundby}{Rounding of the halflife (if the isotope is unstable). Defaults to 1 decimal.}

\item{talk}{Report the results. Defaults to TRUE.}
}
\value{
Invisibly returns a list with the atom's name, symbol, abundance, half-life and decay mode.
A message describing the isotope is printed.
}
\description{
Build an atom
From the number of protons and neutrons, report the corresponding element/isotope, its abundance, half-life, decay mode and daughter product.
}
\details{
The number of protons determines the element, and the number of neutrons the isotope. Some combinations of protons and neutrons are impossible (NA), others exist for only a very short time before decaying, and others are stable.

If only the number of protons is provided, a range of possible amounts of neutrons is returned, et vice versa.

Data downloaded from https://www-nds.iaea.org/relnsd/v1/data?fields=ground_states&nuclides=all
}
\examples{
  atom(1,0)
  atom("carbon")
  atom("C")
  atom(6,7)
  atom(6,8)
  atom("lead")
  atom(82, 128)
}
\author{
Maarten Blaauw
}
