% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realms.R
\name{b2ktoC14}
\alias{b2ktoC14}
\title{Find the 14C age and error belonging to a b2k age.}
\usage{
b2ktoC14(
  x,
  cc = 1,
  postbomb = FALSE,
  rule = 1,
  cc.dir = NULL,
  thiscurve = NULL
)
}
\arguments{
\item{x}{The b2k year.}

\item{cc}{calibration curve for C14 (see \code{caldist()}).}

\item{postbomb}{Whether or not to use a postbomb curve (see \code{caldist()}).}

\item{rule}{How should R's approx function deal with extrapolation. If \code{rule=1}, the default, then NAs are returned for such points and if it is 2, the value at the closest data extreme is used.}

\item{cc.dir}{Directory of the calibration curves. Defaults to where the package's files are stored (system.file), but can be set to, e.g., \code{cc.dir="curves"}.}

\item{thiscurve}{As an alternative to providing cc and/or postbomb, the data of a specific curve can be provided (3 columns: cal BP, C14 age, error).}
}
\value{
The calibration-curve 14C year belonging to the entered b2k age
}
\description{
Given a b2k age (years before AD 2000, popular in the ice core community), the calibration curve (default cc=1) is interpolated and the corresponding 14C age and error are returned.
}
\details{
Interpolation is used, and values outside the calibration curve are given as NA. For ages younger than 50 b2k, a postbomb curve will have to be provided.
}
\examples{
  b2ktoC14(100)
}
\author{
Maarten Blaauw
}
