% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescales.R
\name{b2ktocalBP}
\alias{b2ktocalBP}
\title{calculate cal BP ages from b2k ages}
\usage{
b2ktocalBP(x)
}
\arguments{
\item{x}{The b2k age(s) to be translated into cal BP age(s).}
}
\value{
The cal BP age(s).
}
\description{
calculate cal BP ages from b2k ages
}
\details{
Turn b2k ages (often used in the ice core community, AD 2000) into cal BP ages.
}
\examples{
 b2ktocalBP(0)
}
