% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burden_Scores.R
\name{Burden_Scores}
\alias{Burden_Scores}
\title{Extract Burden Scores from a aGDS}
\usage{
Burden_Scores(
  region = c("Coding", "Noncoding"),
  chr,
  gene_name,
  category = c("plof", "plof_ds", "missense", "disruptive_missense", "synonymous",
    "downstream", "upstream", "UTR", "promoter_CAGE", "promoter_DHS", "enhancer_CAGE",
    "enhancer_DHS", "ncRNA"),
  genofile,
  IDs,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2,
  QC_label = "annotation/filter",
  variant_type = c("SNV", "Indel", "variant"),
  geno_missing_imputation = c("mean", "minor"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  silent = FALSE
)
}
\arguments{
\item{region}{A character value specifying the region of rare variants, either Coding or Noncoding. Coding region contains 5 functional categories \code{plof}, \code{plof_ds}, \code{missense}, \code{disruptive_missense}, or \code{synonymous} and noncoding region contains 8 functional categories \code{downstream}, \code{upstream}, \code{UTR}, \code{promoter_CAGE}, \code{promoter_DHS}, \code{enhancer_CAGE}, \code{enhancer_DHS}, or \code{ncRNA}.}

\item{chr}{A numeric value specifying chromosome in which the protein-coding gene lies in.}

\item{gene_name}{A character value specifying the protein-coding gene.}

\item{category}{A character value specifying the functional category in the protein-coding gene for which the burden score will be constructed from. When region is specified as Coding, choices include \code{plof}, \code{plof_ds}, \code{missense}, \code{disruptive_missense}, or \code{synonymous} and when region is specified as Noncoding choices include \code{downstream}, \code{upstream}, \code{UTR}, \code{promoter_CAGE}, \code{promoter_DHS}, \code{enhancer_CAGE}, \code{enhancer_DHS}, or \code{ncRNA}.}

\item{genofile}{The opened AGDS file using SeqOpen in SeqArray.}

\item{IDs}{Subject IDs of individuals to construct a burden score with. Should either be a character or numeric vector matching the type from SeqGetData(genofile,"sample.id").}

\item{rare_maf_cutoff}{The maximum allele frequency for defining rare variants to be included in the construction of the burden score (default = 0.01).}

\item{rv_num_cutoff}{A numeric value specifying the cutoff of minimum number of variants to include in the construction of a burden score (default = 2).}

\item{QC_label}{Character value specifying the channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{Character value specifying the type of variant included in the analysis. Choices include "SNV", "Indel", or "variant" (default = "SNV").}

\item{geno_missing_imputation}{Character value specifying the method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{Annotation_dir}{Character value specifying the channel name of the annotations in the aGDS file (default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{A data frame containing the name and the corresponding channel name in the aGDS file.}

\item{silent}{Logical: should the report of error messages be suppressed (default = FALSE).}
}
\value{
A column vector containing the burden score for a given chromosome, protein-coding gene, and functional category for rare variants in either the coding or noncoding region.
}
\description{
Extracts a non-weighted Burden Score; b = G x 1, for a given chromosome, protein-coding gene, and functional category for rare variants in either the coding region or noncoding region.
}
