% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\name{C14.cycle}
\alias{C14.cycle}
\title{Fill the atmosphere and ocean with 14C}
\usage{
C14.cycle(
  rate = 6,
  duration = 50000,
  n.steps = 1e+05,
  halflife = 5730,
  CO2.ppm.atm = 300,
  CO2.ppm.ocean = 300,
  mass.atm = 5e+18,
  mass.ocean = 1.4e+21,
  exchange = 0.0195,
  f.ocean = 0.39,
  as.ratio = TRUE,
  C14.1950 = 1.176e-12,
  col.atmosphere = "orange",
  col.ocean = "blue",
  col.rate = "darkgrey",
  bg.rate = rgb(0, 0, 0, 0.05),
  x.lim = c(),
  y.lim = c()
)
}
\arguments{
\item{rate}{Annual production of 14C, in kg. Defaults to 6 kg/yr (see details). Instead of a constant, can also be provided as a time-series (e.g., a random walk).}

\item{duration}{Amount of time over which to cycle, in years.}

\item{n.steps}{Number of steps.}

\item{halflife}{Half-life of 14C. Defaults to Libby Halflife of 5730 yr, but can also be set to much shorter or longer to see the impacts on concentrations.}

\item{CO2.ppm.atm}{Concentration (in ppm) of atmospheric 12C (as CO2). Defaults to 300 ppm.}

\item{CO2.ppm.ocean}{Concentration (in ppm) of surface ocean 12C. Defaults to 300 ppm.}

\item{mass.atm}{Mass (in kg) of carbon in the atmosphere (as CO2).}

\item{mass.ocean}{Mass (in kg) of carbon in the ocean.}

\item{exchange}{Exchange rate of carbon (as CO2) between the atmosphere and the ocean. Defaults to 0.0195 (since this results in reasonable atmospheric 14C concentrations).}

\item{f.ocean}{Fraction of the ocean partaking in exchange with the atmosphere. Implicitly models surface and deep ocean. Defaults to 0.39.}

\item{as.ratio}{Plot carbon in the reservoirs as 14C/12C ratio (default). If set to FALSE, is plotted as mass of 14C.}

\item{C14.1950}{The standard 14C/C ratio at 0 cal BP (AD 1950), defined as 95\% of the C-14 activity of NBS Oxalic Acid I (oxI), normalized to d13C=–25 permille. Set to NA to avoid plotting the dashed horizontal line.}

\item{col.atmosphere}{Colour of the curve depicting the atmosphere's values. Defaults to sunny orange.}

\item{col.ocean}{Colour of the curve depicting the ocean's values. Defaults to blue.}

\item{col.rate}{Colour of the rate (annual addition of C14 to the atmosphere). Defaults to dark grey.}

\item{bg.rate}{Colour of the panel behind the rate curve (annual addition of C14 to the atmosphere). Defaults to light grey.}

\item{x.lim}{Axis limits of the horizontal axis. Calculated automatically by default.}

\item{y.lim}{Axis limits of the vertical axis. Calculated automatically by default.}
}
\value{
A plot and the underlying values of the atmospheric and ocean 14C concentration (in kg and as proportion of 12C).
}
\description{
Simulates the accumulation and exchange of 14C within the atmosphere and ocean over time, based on annual production, decay and exchange between the atmosphere and ocean.
}
\details{
This is a very simplistic 'toy' model with two layers (atmosphere and ocean) between which carbon is exchanged. The model ignores essential carbon cycling features such as vegetation, soils, ocean layers, circulation and upwelling and should not be taken as anything other than an educational tool. 

The simulation starts with no C14 present in either of the reservoirs (but both reservoirs will have plenty C12). As time passes, C14 produced by cosmic rays accumulates inside the atmosphere, with some leaking to the ocean at each step. Also at each step, some of the C14 will decay (at a rate dictated by its halflife). At each step, the total mass of 14C in both reservoirs is returned as well as its proportion as 14C/12C. Users can play around with values of annual 14C production, step duration, total mass, 12C concentration, exchange between the atmosphere and ocean, and half-life.

Estimates of how much C14 is produced per year vary, but can be approached as follows. Up in the atmosphere (production peaks at c. 10 km height, where airplanes fly), cosmic ray collisions produce c. 1-2 14C atoms per second per cm^2 on average (c. 17k/m2; actual rates vary and are higher near the poles). Since the surface of a sphere equals 4*pi*r^2 and given Earth's radius of 6371 km, the amount of C14 particles produced per year would be n = 17e3 * 365.25 * 24*60*60 * 4*pi*(1e3*(6371+10))^2 or c. 2.745e26 C14 particles. Multiplied by the mass of a single C14 atom, n * 2.32e-26 = c. 6.37 kg. We round this to 6 kg, about the weight of a well-fed (i.e., slightly chonky) cat. 

Default values for the total masses of the atmosphere and the ocean are from Wikipedia, and C-12 concentrations are set at 300 ppm (i.e., pre-industrial atmospheric CO2 concentrations). With these values and the above C-14 production rate, the atmosphere contains around 480 kg of C14 at equilibrium, about the mass of a polar bear. A further 50,000 kg of C-14 resides in the ocean, roughly the mass of 10 African Elephants. The coefficient that sets the rate of exchange between the atmosphere and ocean is tuned such that with default settings, 14C/12C ratios reach expected values of c. 1.2e-12.
}
\examples{
  C14.cycle()
  mn <- 6; rw <- mn; set.seed(67)
  for(i in 2:1e5) rw[i] <- rw[i-1] + rnorm(1, 0, .001)
  C14.cycle(rw)
}
\author{
Maarten Blaauw
}
