% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realms.R
\name{calBPtob2k}
\alias{calBPtob2k}
\title{calculate b2k ages from cal BP ages}
\usage{
calBPtob2k(x)
}
\arguments{
\item{x}{The calBP age(s) to be translated into b2k ages.}
}
\value{
The b2k ages.
}
\description{
calculate b2k ages from cal BP ages
}
\details{
Turn cal BP ages into b2k ages (years before AD 2000), which are often used in the ice core community.
}
\examples{
 calBPtob2k(-50)
}
