% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\name{decay_series}
\alias{decay_series}
\title{Decay series of isotopes}
\usage{
decay_series(
  protons,
  neutrons,
  plot = TRUE,
  add = FALSE,
  n = c(),
  box.size = 0.8,
  arrows = TRUE,
  draw.box = TRUE,
  draw.symbol = TRUE,
  grid.col = grey(0.75),
  xlim = c(),
  ylim = c(),
  cex = 2,
  offset = 0.5,
  mar = c(3, 3, 0.5, 0.5),
  mgp = c(1.5, 0.7, 0),
  asp = 0,
  talk = TRUE
)
}
\arguments{
\item{protons}{The number of protons in the atom's nucleus. Alternatively, the name or symbol of an element can provided.}

\item{neutrons}{The number of neutrons in the atom's nucleus. If left empty, the function will report the amounts of neutrons of the isotopes of the element in question.}

\item{plot}{Plot the results. Defaults to TRUE.}

\item{add}{Add data to an existing plot (instead of making a new plot). Defaults to FALSE.}

\item{n}{Number of isotopes to draw, starting from the first isotope. Defaults to drawing all isotopes in the decay series.}

\item{box.size}{Size of the box around the isotope symbols. Defaults to 0.9.}

\item{arrows}{Whether or not to draw arrows between the parent and daughter isotopes. Defaults to TRUE.}

\item{draw.box}{Whether or not to draw a box around the element. Defaults to TRUE.}

\item{draw.symbol}{Whether or not to draw the element's symbol, e.g. 14C. Defaults to TRUE.}

\item{grid.col}{The colour of the grid. Defaults to light grey, \code{grid.col=grey(.75)}.}

\item{xlim}{Limits of the x axis. Calculated automatically by default.}

\item{ylim}{Limits of the y axis. Calculated automatically by default.}

\item{cex}{Controls how much of each isotope box is filled by the isotope's symbol (with superscript mass numbers). Values around 2–3 typically produce labels that occupy most of the box while remaining legible. Defaults to 2.}

\item{offset}{Offset of the symbol within the isotope's box. Defaults to centre, 0.5.}

\item{mar}{Margins around the plot. Defaults to mar=c(3,3,0.5,0.5)}

\item{mgp}{Locations of the axis titles, labels and lines. Defaults to mgp=(1.5, 0.7, 0)}

\item{asp}{If set to 1, plots the units of the horizontal and vertical axis at 1:1 proportion.}

\item{talk}{Report the results. Defaults to TRUE.}
}
\description{
Given a starting isotope, show the decay chain all the way to reaching a stable isotope.
}
\details{
The `atom` function is called for each of the isotopes in the decay series, and a plot is made.
}
\examples{
  decay_series("U", 146) # U-238 series
  decay_series("U", 143) # U-235 series
  decay_series("Th", 142) # Thorium series
}
\author{
Maarten Blaauw
}
