% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elements}
\alias{elements}
\title{elements Data}
\format{
A list with 7 variables and 282 entries for each variable.
\describe{
  \item{protons}{amount of protons in the nucleus}
  \item{neutrons}{amount of neutrons in the nucleus}
  \item{name}{name of the element}
  \item{symbol}{symbol of the element}
  \item{decay}{type of decay product (if not stable)}
  \item{halflife}{decay half-life}
  \item{abundance}{abundance of the isotope, relative to the element}
}
}
\source{
Data downloaded from https://www-nds.iaea.org/relnsd/v1/data?fields=ground_states&nuclides=all
}
\usage{
elements
}
\description{
A dataset containing the amounts of protons and neutrons, names, symbols, decay type and halflife for all naturally occurring isotopes
}
\examples{
data(elements)
head(elements)
}
\keyword{datasets}
