% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensemble_Regression.R
\name{Ensemble_Function}
\alias{Ensemble_Function}
\title{Perform Rare Variant PRS Ensemble Regression}
\usage{
Ensemble_Function(PRSs, Y, family = c("continuous", "binary"))
}
\arguments{
\item{PRSs}{A data.frame or matrix containing p PRSs and n observations. p must be greater than 1 and n must be equal the number of observations in Y. PRSs and Y must be matched before performing ensemble regression.}

\item{Y}{A data.frame, matrix, or vector containing continuous or binary outcome Y. Number of observations must equal the number of observations in PRSs. PRSs and Y must be matched before performing ensemble regression.}

\item{family}{Either continuous or binary depending on outcome Y. Defaults to continuous.}
}
\value{
A named list with one item, Coefficients. Coefficients contain coefficients of the PRSs columns to generate the predicted ensemble PRS. Coefficients include the intercept estimate.
}
\description{
Performs ensemble regression of PRSs and continuous or binary response Y. Produces a set of coefficients for the PRSs.
}
