% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realms.R
\name{F14CtoC14}
\alias{F14CtoC14}
\title{Calculate C14 ages from F14C values.}
\usage{
F14CtoC14(F14C, er = NULL, decimals = 8, lambda = 8033)
}
\arguments{
\item{F14C}{Reported mean of the F14C}

\item{er}{Reported error of the F14C. Returns just the mean if left empty.}

\item{decimals}{Amount of decimals required for the radiocarbon age. Quite sensitive, defaults to 8.}

\item{lambda}{The mean-life of radiocarbon (based on Libby half-life of 5568 years)}
}
\value{
The radiocarbon ages from the F14C values. If F14C values are above 100\%, the resulting radiocarbon ages will be negative.
}
\description{
Calculate C14 ages from F14C values of radiocarbon dates.
}
\details{
Post-bomb dates are often reported as F14C (between 0 at c. 55 kcal BP and 1 at c. AD 1950). Since software such as Bacon expects radiocarbon ages,
 this function can be used to calculate radiocarbon ages from F14C values. The reverse function is \link{age.F14C}.
}
\examples{
  F14CtoC14(1.10, 0.005) # a postbomb date, so with a negative C14 age
  F14CtoC14(.80, 0.005) # prebomb dates can also be calculated
}
