% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescales.R
\name{F14CtoDelta14C}
\alias{F14CtoDelta14C}
\title{Transform F14C into Delta14C}
\usage{
F14CtoDelta14C(F14C, er = NULL, t, roundby = Inf, lambda = 5730/log(2))
}
\arguments{
\item{F14C}{The F14C value to translate}

\item{er}{Reported error of the F14C. Returns just the mean if left empty.}

\item{t}{the cal BP age}

\item{roundby}{Amount of decimals required for the output. Defaults to \code{roundby=Inf}, no rounding.}

\item{lambda}{Radiocarbon's mean-life, based on the Cambridge half-life}
}
\value{
The corresponding Delta14C value
}
\description{
Transform F14C into Delta14C
}
\details{
As explained by Heaton et al. 2020 (Radiocarbon), 14C measurements are commonly expressed in
three domains: Delta14C, F14C and the radiocarbon age. This function translates F14C values into Delta14C, the historical level of Delta14C in the year t cal BP. Note that per convention, this function uses the Cambridge half-life, not the Libby half-life.
}
\examples{
  F14CtoDelta14C(0.89, .001, 900)
}
