% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescales.R
\name{fromto}
\alias{fromto}
\title{translate between timescales}
\usage{
fromto(
  x,
  from = "calBP",
  cc = 1,
  postbomb = 1,
  cc.dir = NULL,
  thiscurve = NULL,
  zero = FALSE,
  width = c(),
  digits = 0,
  C14.col = rgb(0, 0, 1, 0.5),
  Delta14C.col = rgb(0, 0.4, 0, 0.4),
  ka = FALSE,
  cal.rev = TRUE,
  legend.size = 0.7
)
}
\arguments{
\item{x}{The value to be translated into other timescales}

\item{from}{The timescales of the entered value. Can be "calBP" for cal BP, "BCAD" for BC/AD, "C14" for C14 BP, "F14C" for F14C, or "pMC" for pMC. Delta14C cannot be entered as a value (you could enter the corresponding cal BP or BC/AD ages instead).}

\item{cc}{calibration curve for C14 (see \code{caldist()}).}

\item{postbomb}{Whether or not to use a postbomb curve (see \code{caldist()}).}

\item{cc.dir}{Directory of the calibration curves. Defaults to where the package's files are stored (system.file), but can be set to, e.g., \code{cc.dir="curves"}.}

\item{thiscurve}{As an alternative to providing cc and/or postbomb, the data of a specific curve can be provided (3 columns: cal BP, C14 age, error).}

\item{zero}{Whether or not zero BC/AD should be included. Defaults to none, FALSE.}

\item{width}{Width of the righthand plot. Calculated automatically by default (older ages get wider windows).}

\item{digits}{Rounding of the reported values. Defaults to 0 digits.}

\item{C14.col}{Colour of the 14C calibration curve. Defaults to semi-transparent blue, \code{C14.col=rgb(0,0,1,.5)}.}

\item{Delta14C.col}{Colour of the Delta14C curve. Defaults to semi-transparent green, \code{Delta14C.col=rgb(0,.4,0,.4)}.}

\item{ka}{Whether to use years or ka (thousands of years). Defaults to \code{ka=FALSE}.}

\item{cal.rev}{Reverse the age axis (right panel). Defaults to TRUE.}

\item{legend.size}{Size of the font of the legend. Defaults to 0.7 of R's standard size.}
}
\value{
A plot and output showing the translations into the different timescales.
}
\description{
translate between timescales
}
\details{
Upon entering a value and its timescale, this function will find the corresponding values in the other timescales. Note that uncertainties are *not* taken into account, and especially going from C14 BP to cal BP and BC/AD ignores many calibration-related uncertainties. Delta14C values are only reported for entered values on the cal BP or BC/AD scale.
}
\examples{
  fromto(0, "BCAD")
  fromto(2450, "C14")
}
