% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\name{howmanyC14}
\alias{howmanyC14}
\title{Amount of C14 particles in a sample}
\usage{
howmanyC14(
  age,
  wght = 1,
  use.cc = TRUE,
  Av = 6.02214076e+23,
  C14.ratio = 1.176e-12,
  format = "g",
  cc = 1,
  postbomb = FALSE,
  cc.dir = NULL,
  thiscurve = NULL,
  talk = TRUE,
  decimals = 4
)
}
\arguments{
\item{age}{The age of the sample (in cal BP per default, or in C14 BP is use.cc=FALSE).}

\item{wght}{The weight of the sample (in mg). Defaults to 1 mg.}

\item{use.cc}{Whether or not to use the calibration curve. If set to \code{use.cc=FALSE}, then we assume that the age is the radiocarbon age (this enables ages beyond the reach of the calibration curves to be used).}

\item{Av}{Avogadro's number, used to calculate the number of carbon atoms in the sample.}

\item{C14.ratio}{The 14C/C ratio at F=1 (AD 1950).}

\item{format}{The format of the printed numbers. Defaults to either scientific (for large numbers) or as fixed-point, depending on the size of the number.}

\item{cc}{calibration curve for C14 (see \code{caldist()}).}

\item{postbomb}{Whether or not to use a postbomb curve (see \code{caldist()}).}

\item{cc.dir}{Directory of the calibration curves. Defaults to where the package's files are stored (system.file), but can be set to, e.g., \code{cc.dir="curves"}.}

\item{thiscurve}{As an alternative to providing cc and/or postbomb, the data of a specific curve can be provided (3 columns: cal BP, C14 age, error).}

\item{talk}{Whether or not to provide feedback (defaults to TRUE).}

\item{decimals}{Number of decimals to be returned for F and atom counts.}
}
\value{
The estimated number of C14 atoms.
}
\description{
Find the amount of remaining C14 atoms in a sample, given its weight and age.
}
\details{
The number of carbon atoms in the sample is estimated. Given the known C14/C ratio at F=1, and given the sample's age, we can estimate the number of remaining C14 atoms.
}
\examples{
  calBPtoD14C(100)
}
\author{
Maarten Blaauw
}
