% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\name{map.dates}
\alias{map.dates}
\title{A map of 180k archaeological C-14 dates}
\usage{
map.dates(
  S = 48,
  W = -15,
  N = 62,
  E = 5,
  download = FALSE,
  rainbow = FALSE,
  mincol = "yellow",
  maxcol = "red",
  size = 1.5,
  legend.loc = "topright"
)
}
\arguments{
\item{S}{The southern limit of the initial map.}

\item{W}{The western limit of the initial map.}

\item{N}{The northern limit of the initial map.}

\item{E}{The eastern limit of initial map.}

\item{download}{Whether or not to try to download the p13k14c data. If set to FALSE (default option) and if the p3k14c package is not installed, instructions to do so are provided. If TRUE and p3k14c is not installed, the relevant file will be downloaded from `www.p3k14c.org`. Note: it weighs around 26 MB (c. 4 MB zipped).}

\item{rainbow}{Whether or not to use a rainbow scale to plot the variable.}

\item{mincol}{Colour for minimum values. Defaults to 'yellow'.}

\item{maxcol}{Colour for maximum values. Defaults to 'red'.}

\item{size}{Size of the symbols. Defaults to 1.5.}

\item{legend.loc}{Location of the legend. Defaults to the top right.
## Not run: 
  alldates <- map.dates()}
}
\value{
An interactive, browseable map returning all radiocarbon dates (age and Lab ID) in the p3k14c database.
}
\description{
Produce an interactive, browseable map of the c. 180k (!) archaeological radiocarbon dates provided by the p3k14c R package.
}
\details{
The p3k14c R package is available on github ("people3k/p3k14c"). See Bocinsky, R. Kyle, Darcy Bird, Lux Miranda, and Jacob Freeman (2022.6). Compendium of R code and data for p3k14c: A synthetic global database of archaeological radiocarbon dates. Accessed 1 July 2025. https://doi.org/10.5281/zenodo.6633635. The p3k14c data has been provided under a CC-0 license.
Bird, D., Miranda, L., Vander Linden, M. et al. p3k14c, a synthetic global database of archaeological radiocarbon dates. Sci Data 9, 27 (2022). https://doi.org/10.1038/s41597-022-01118-7
You are recommended to have a look at the github instructions for the p3k14c package. The `map.dates` function is provided here for an initial visual analysis of the data; p3k14c provides functions to get much more out of the data.
Plotting requires the 'leaflet' R package to be installed. An error will occur if it isn't. If your version of R is below 4.0.0 (released back in April 2020), you're prompted to install a more recent version.
}
