% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realms.R
\name{pMC.age}
\alias{pMC.age}
\title{To be deprecated. Use pMCtoC14 instead.}
\usage{
pMC.age(mn, sdev = c(), ratio = 100, decimals = 0, lambda = 8033)
}
\arguments{
\item{mn}{Reported mean of the pMC.}

\item{sdev}{Reported error of the pMC.}

\item{ratio}{Most modern-date values are reported against \code{100}. If it is against \code{1} instead, use \code{1} here.}

\item{decimals}{Amount of decimals required for the radiocarbon age.}

\item{lambda}{The mean-life of radiocarbon (based on Libby half-life of 5568 years)}
}
\value{
Radiocarbon ages from pMC values. If pMC values are above 100\%, the resulting radiocarbon ages will be negative.
}
\description{
Will be deprecated. Use pMCtoC14 instead.
}
\details{
Post-bomb dates are often reported as pMC or percent modern carbon. Since Bacon expects radiocarbon ages,
 this function can be used to calculate radiocarbon ages from pMC values. The reverse function is C14.pMC.
}
