% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\name{radio}
\alias{radio}
\title{Listen to radiocarbon being measured}
\usage{
radio(
  age,
  duration = 10,
  sr = 44100,
  as.clicks = TRUE,
  as.tone = TRUE,
  click_length = 80,
  tone.volume = 0.5,
  noise = 0.02,
  meander = 0.005,
  play = interactive(),
  ...
)
}
\arguments{
\item{age}{The age of the sample (in cal BP per default, or in C14 BP if use.cc=FALSE).}

\item{duration}{How long the sample will sound for in seconds. Defaults to 10 seconds.}

\item{sr}{Sampling rate. This defaults to 44100 (per second), which is based on the CD standard.}

\item{as.clicks}{Make the C-14 counts sound as clicks, based on random Poisson sampling.}

\item{as.tone}{Make the C-14 frequency (counts per second) sound as a wave (e.g., 105 cps becomes a 105 Hz sine wave). This can either be a constant wave or be meandering (see `noise` and `meander`).}

\item{click_length}{Length of the clicks. Defaults to 80, but can be altered to make the clicks sound different}

\item{tone.volume}{Volume of the tone/wave relative to that of the clicks.}

\item{noise}{Deviation from the mean of the tone. Defaults to 0.02.}

\item{meander}{Drift of the tone along the mean. Defaults to 0.005.}

\item{play}{Whether or not to play the sound}

\item{...}{Optional constants to be entered into the function `howmuchC14`}
}
\value{
A sound (tone and or clicks) is played and returned invisibly.
}
\description{
A sonification of the amount of C14 being detected in an AMS.
}
\details{
The sonification can be heard as 'Geiger counter-like' clicks (individual C14 atoms ending up in the detector) and/or as a possibly slightly meandering sine wave (e.g., for a sample of age 5000 14C BP, we'd count c. 105.6 atoms per second in the AMS, so this would become a tone of frequency 105.6 Hz). The calculations are based on the function \code{howmuchC14}.
}
\examples{
  radio(0)
  radio(45000)
}
\author{
Maarten Blaauw
}
