% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice-package.R
\docType{package}
\name{rice-package}
\alias{rice-package}
\title{rice: Radiocarbon Equations}
\description{
Provides functions for the calibration of radiocarbon dates, as well as options to calculate different radiocarbon timescales (C14 age, F14C, pMC, D14C) and estimating the effects of contamination or local reservoir offsets (Reimer and Reimer 2001 \doi{10.1017/S0033822200038339}). The methods follow long-established recommendations such as Stuiver and Polach (1977) \doi{10.1017/S0033822200003672} and Reimer et al. (2004) \doi{10.1017/S0033822200033154}. This package uses the calibration curves from the data package 'rintcal'.
}
\author{
\strong{Maintainer}: Maarten Blaauw \email{maarten.blaauw@qub.ac.uk} (\href{https://orcid.org/0000-0002-5680-1515}{ORCID})

}
