% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shroud.R
\docType{data}
\name{shroud}
\alias{shroud}
\title{shroud Data}
\format{
A data frame with the Lab numbers, ages and errors of 12 Shroud dates.
\describe{
  \item{ID}{Lab numbers. Replicates are indicates with .1, .2, etc.}
  \item{y}{Radiocarbon year}
  \item{er}{Lab error}
}
}
\source{
Data taken from Damon et al. 1989 [Nature] <doi:10.1038/337611a0>, see also Christen 1994 [Applied Statistics] <doi:10.2307/2986273>
}
\usage{
shroud
}
\description{
A dataset containing the radiocarbon dates on the Shroud of Turin, from three labs
}
\examples{
data(shroud)
head(shroud)
}
\keyword{datasets}
