% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{sim.date}
\alias{sim.date}
\title{Simulate a radiocarbon date}
\usage{
sim.date(
  x,
  n = 1,
  F.er = 0.002,
  scatter = 1,
  error.multiplier = 1,
  cc.error = TRUE,
  cc = 1,
  postbomb = FALSE,
  glue = 0,
  BCAD = FALSE,
  zero = FALSE,
  thiscurve = NULL,
  cc.dir = NULL,
  rule = 1,
  deltaR = 0,
  deltaSTD = 0,
  round = 0
)
}
\arguments{
\item{x}{The calendar age (in cal BP by default)}

\item{n}{Number of dates to simulate. Note that whereas a real radiocarbon date costs several hundreds of GBP/USD/EUR and takes months to be reported, there's no need to pay or wait while using \code{simulate.date} and you could set n to, say, 100 at no extra cost. Defaults to just 1 radiocarbon date.}

\item{F.er}{The laboratory error of the radiocarbon date, on the F scale. Defaults to 2 permille.}

\item{scatter}{Randomness of the dates. If scatter > 0, dates are sampled from a normal distribution with standard deviation = scatter * total uncertainty. Set scatter=0 to return the expected radiocarbon age without random variation.}

\item{error.multiplier}{Multiplier of the laboratory uncertainty (F.er). Some radiocarbon labs have different error multipliers for different materials. Defaults to 1.}

\item{cc.error}{Whether calibration-curve uncertainties should be included. Defaults to TRUE. Setting this to FALSE treats the calibration curve as known without uncertainty/error and is intended primarily for simulations and sensitivity analyses.}

\item{cc}{The calibration curve to smooth. Calibration curve for 14C dates: 'cc=1' for IntCal20 (northern hemisphere terrestrial), 'cc=2' for Marine20 (marine), 'cc=3' for SHCal20 (southern hemisphere terrestrial). Alternatively, one can also write, e.g., "IntCal20", "Marine13". One can also make a custom-built calibration curve, e.g. using 'mix.ccurves()', and load this using 'cc=4'. In this case, it is recommended to place the custom calibration curve in its own directory, using 'cc.dir' (see below). Explanations of the numbers are provided in the table footer. If there is more than one cc provided, they will be printed in an extra table column.}

\item{postbomb}{Use 'postbomb=TRUE' to get a postbomb calibration curve (default 'postbomb=FALSE'). For monthly data, type e.g. 'cc="sh1-2_monthly"'}

\item{glue}{Glue postbomb and prebomb curves together. Defaults to 0 (none), can be 1 (IntCal20 + NH1), 2 (IntCal20 + NH2), 3 (IntCal20 + NH3), 4 (SHCal20 + SH1-2) or 5 (SHCal20 + SH3). Note that this will override the value of cc.}

\item{BCAD}{Which calendar scale to use. Defaults to cal BP, \code{BCAD=FALSE}. For the BCAD scale, BC ages are negative.}

\item{zero}{Whether or not zero BC/AD should be included if using BCAD=TRUE. Defaults to \code{zero=FALSE}.}

\item{thiscurve}{As an alternative to providing cc and/or postbomb, the data of a specific curve can be provided (3 columns: cal BP, C14 age, error). Defaults to c().}

\item{cc.dir}{Directory of the calibration curves. Defaults to where the package's files are stored (system.file), but can be set to, e.g., 'cc.dir="ccurves"'.}

\item{rule}{Approximation rule for finding the calibration curve's C14 age.}

\item{deltaR}{Age offset (e.g. for marine samples). If present, the age offset will be simulated as a normal distribution, deltaR +-deltaSTD.}

\item{deltaSTD}{Uncertainty of the age offset (1 standard deviation). If present, the age offset will be simulated as a normal distribution, deltaR +-deltaSTD.}

\item{round}{Rounding for the output. Defaults to 0 decimals. To round to the year, use \code{round=0}, for the nearest decade, use \code{round=-1}, for centuries use \code{round=-2}, for near-monthly resolution use \code{round=1}, and so on. Set to \code{round=Inf} to retrieve all digits.}
}
\value{
The simulated C14 age and error
}
\description{
Simulate a radiocarbon date based on a calendar age (cal BP or BC/AD), the corresponding calibration curve's C14 age, scatter and any offset.
}
\details{
The calibration curve is queried at the requested cal BP age(s). Laboratory uncertainty, calibration-curve uncertainty and any offset uncertainty are combined in quadrature.
}
\examples{
 sim.date(900, 5, BCAD=TRUE)
}
\author{
Maarten Blaauw
}
