% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{younger}
\alias{younger}
\title{Find the probability of a calibrated distribution being of a certain age or younger than it}
\usage{
younger(
  x,
  y,
  er = c(),
  cc = 1,
  postbomb = FALSE,
  glue = 0,
  bombalert = TRUE,
  deltaR = 0,
  deltaSTD = 0,
  normal = TRUE,
  as.F = FALSE,
  is.F = FALSE,
  t.a = 3,
  t.b = 4,
  BCAD = FALSE,
  zero = FALSE,
  threshold = 0
)
}
\arguments{
\item{x}{The year or years of interest, in cal BP by default.}

\item{y}{The radiocarbon date's mean. Alternatively, can be an age distribution (two columns: ages and their probabilities) or a vector of ages.}

\item{er}{The radiocarbon date's lab error. Leave empty if y is an age distribution.}

\item{cc}{calibration curve for the radiocarbon date(s) (see the \code{rintcal} package).}

\item{postbomb}{Whether or not to use a postbomb curve (see \code{caldist()}).}

\item{glue}{Glue postbomb and prebomb curves together. Defaults to 0 (none), can be 1 (IntCal20 + NH1), 2 (IntCal20 + NH2), 3 (IntCal20 + NH3), 4 (SHCal20 + SH1-2) or 5 (SHCal20 + SH3). Note that this will override the value of cc.}

\item{bombalert}{Warn if a date is close to the lower limit of the calibration curve. Defaults to \code{postbomb=TRUE}.}

\item{deltaR}{Age offset (e.g. for marine samples).}

\item{deltaSTD}{Uncertainty of the age offset (1 standard deviation).}

\item{normal}{Use the normal distribution to calibrate dates (default TRUE). The alternative is to use the t model (Christen and Perez 2016).}

\item{as.F}{Whether or not to calculate ages in the F14C timescale. Defaults to \code{as.F=FALSE}, which uses the C14 timescale.}

\item{is.F}{Use this if the provided date is in the F14C timescale.}

\item{t.a}{Value a of the t distribution (defaults to 3).}

\item{t.b}{Value b of the t distribution (defaults to 4).}

\item{BCAD}{Which calendar scale to use. Defaults to cal BP, \code{BCAD=FALSE}.}

\item{zero}{Whether or not zero BC/AD should be included if using BCAD=TRUE. Defaults to \code{zero=FALSE}.}

\item{threshold}{Report only values above a threshold. Defaults to \code{threshold=0}.}
}
\value{
The probability of a date being of a certain calendar age or younger than it. Returns a vector if x contains multiple ages.
}
\description{
Find the probability that a date or distribution is of a certain calendar age x or younger than it, by calculating the proportion of the calibrated distribution up to and including x (i.e., summing the calibrated distribution up to year x).
}
\details{
Instead of providing a to-be-calibrated age (y +- er), one can also provide an age distribution (as two columns: cal BP ages and their probabilities) or a vector of ages.
}
\examples{
younger(2800, 2450, 20)
younger(2400, 2450, 20)
calibrate(160, 20, BCAD=TRUE)
younger(1750, 160, 20, BCAD=TRUE)
}
\author{
Maarten Blaauw
}
