% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realms.R
\name{F14CtopMC}
\alias{F14CtopMC}
\title{Calculate pMC ages from F14C values.}
\usage{
F14CtopMC(mn, sdev = NULL)
}
\arguments{
\item{mn}{Reported mean of the F14C}

\item{sdev}{Reported error of the F14C. Returns just the mean if left empty.}
}
\value{
The pMC values from the F14C values. Basically the original values multiplied by 100.
}
\description{
Calculate pMC vlaues from F14C values of radiocarbon dates.
}
\details{
Post-bomb dates are often reported as F14C (between 0 at c. 55 kcal BP and 1 at c. AD 1950). Since software such as Bacon expects radiocarbon ages,
 this function can be used to calculate radiocarbon ages from F14C values. The reverse function is \link{age.F14C}.
}
\examples{
  F14CtopMC(1.10, 0.5)
}
