% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\name{map.dates}
\alias{map.dates}
\title{A browseable map of 180k archaeological C-14 dates}
\usage{
map.dates(
  S = 48,
  W = -15,
  N = 62,
  E = 5,
  fl = "~/Dropbox/devsoftware/p3k14c_2022.06.csv",
  download = FALSE,
  rainbow = FALSE,
  mincol = "yellow",
  maxcol = "red",
  size = 2,
  legend.loc = "topright"
)
}
\arguments{
\item{S}{The southern limit of the rectangular region.}

\item{W}{The western limit of the rectangular region.}

\item{N}{The northern limit of the rectangular region.}

\item{E}{The eastern limit of the rectangular region.}

\item{fl}{The file containing the database. You can download this yourself from 'https://www.p3k14c.org/download/'.}

\item{download}{Whether or not to try to download the database .csv file. Note: it weighs around 26 MB (c. 4 MB zipped).}

\item{rainbow}{Whether or not to use a rainbow scale to plot the variable.}

\item{mincol}{Colour for minimum values. Defaults to 'yellow'.}

\item{maxcol}{Colour for maximum values. Defaults to 'red'.}

\item{size}{Size of the symbols. Defaults to 2.}

\item{legend.loc}{Location of the legend, if using a basic plot. Defaults to the top right.
## Not run: 
  alldates <- map.dates()}
}
\value{
An interactive, browseable map returning all radiocarbon dates (age and Lab ID) in the database.
}
\description{
Produce an interactive, browseable map of the c. 180k archaeological radiocarbon dates provided by the p3k14c R package. The p3k14c R package has been removed from CRAN (probably owing to the retirement of other mapping-related packages). Here we download the 2022 version of the database as a .csv file from url 'https://www.p3k14c.org/download/'.
}
\details{
See Bocinsky, R. Kyle, Darcy Bird, Lux Miranda, and Jacob Freeman, (2022.6). Compendium of R code and data for p3k14c: A synthetic global database of archaeological radiocarbon dates. Accessed 1 July 2025. https://doi.org/10.5281/zenodo.6633635
requires the 'leaflet' R package to be installed. An error will occur if it isn't installed.
Bird, D., Miranda, L., Vander Linden, M. et al. p3k14c, a synthetic global database of archaeological radiocarbon dates. Sci Data 9, 27 (2022). https://doi.org/10.1038/s41597-022-01118-7
}
