% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\name{radio}
\alias{radio}
\title{Listen to radiocarbon being measured}
\usage{
radio(
  age,
  duration = 10,
  duration.unit = c(),
  use.cc = FALSE,
  as.decays = FALSE,
  wght = 1,
  as.clicks = TRUE,
  as.tone = TRUE,
  click_length = 80,
  tone.volume = 0.5,
  noise = 0.02,
  meander = 0.005,
  play = interactive(),
  sr = 44100,
  visualise = TRUE,
  cex = 0.6,
  return.sound = FALSE,
  ...
)
}
\arguments{
\item{age}{The age of the sample (in C14 BP by default, or in cal BP if use.cc=FALSE).}

\item{duration}{How long the sample will count (and sound) for in seconds. Defaults to 10 seconds.}

\item{duration.unit}{Unit of the duration. Left empty by default (c()) which then makes seconds the unit ('s'), or if as.decays is used, days ('d'). Even if set to 'd', the sound will be played back using the seconds unit ('s').}

\item{use.cc}{Whether or not to use the calibration curve. If set to \code{use.cc=FALSE} (the default), then we assume that the age is the radiocarbon age (this enables ages beyond the reach of the calibration curves to be used).}

\item{as.decays}{Work with the C14 decays. Defaults to FALSE, which works with the number of C14 atoms instead. If set to true, you'll need larger sample sizes (wght) and longer counting times (duration) to get decent counts.}

\item{wght}{The weight of the sample (in mg). Defaults to 1 mg.}

\item{as.clicks}{Make the C-14 counts sound as clicks, based on random Poisson sampling.}

\item{as.tone}{Make the C-14 frequency (counts per second) sound as a wave (e.g., 105 cps becomes a 105 Hz sine wave). This can either be a constant wave or be meandering (see `noise` and `meander`).}

\item{click_length}{Length of the clicks. Defaults to 80, but can be altered to make the clicks sound different}

\item{tone.volume}{Volume of the tone/wave relative to that of the clicks.}

\item{noise}{Deviation from the mean of the tone. Defaults to 0.02.}

\item{meander}{Drift of the tone along the mean. Defaults to 0.005.}

\item{play}{Whether or not to play the sound}

\item{sr}{Sampling rate. This audio quality defaults to 44100 (per second), which is based on the CD standard.}

\item{visualise}{the counts and calculation of the C14 age. Defaults to TRUE.}

\item{cex}{Size of the font. Defaults to \code{cex=0.6}.}

\item{return.sound}{Return the sound as an invisible object. If set to FALSE (the default), instead returns the counts and the calculated C14 age.}

\item{...}{Optional constants to be entered into the function `howmuchC14`}
}
\value{
A sound (tone and or clicks) is played and returned invisibly.
}
\description{
A sonification of the amount of C14 being detected in an AMS.
}
\details{
The sonification can be heard as 'Geiger counter-like' clicks (individual C14 atoms ending up in the detector) and/or as a possibly slightly meandering sine wave (e.g., for a sample of age 5000 14C BP, we'd count c. 105.6 atoms per second in the AMS, so this would become a tone of frequency 105.6 Hz). The calculations are based on the function \code{howmuchC14}.

Instead of counting the particles themselves (as done in an AMS), we can also count the decay events as used in radiometric counting. In this case, we need much larger samples (wght) and longer counting times (duration).

Note that the calculations in this function do not include error multipliers or corrections such as for fractionation and backgrounds.
}
\examples{
  radio(0)
  radio(45000)
  # decay events over 1 minute in 1 gram of carbon of age 500 14C BP:
  radio(500, wght=1000, as.decays=TRUE, duration=60) 
}
\author{
Maarten Blaauw
}
