% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{younger}
\alias{younger}
\title{Find the probability of a calibrated date being of a certain age or younger than it}
\usage{
younger(x, y, er, cc = 1, normal = TRUE, t.a = 3, t.b = 4, BCAD = FALSE)
}
\arguments{
\item{x}{The year of interest, in cal BP by default.}

\item{y}{The radiocarbon date's mean.}

\item{er}{The radiocarbon date's lab error.}

\item{cc}{calibration curve for the radiocarbon date(s) (see the \code{rintcal} package).}

\item{normal}{Use the normal distribution to calibrate dates (default TRUE). The alternative is to use the t model (Christen and Perez 2016).}

\item{t.a}{Value a of the t distribution (defaults to 3).}

\item{t.b}{Value b of the t distribution (defaults to 4).}

\item{BCAD}{Which calendar scale to use. Defaults to cal BP, \code{BCAD=FALSE}.}
}
\value{
The probability of a date being of a certain calendar age or younger than it.
}
\description{
Find the probability that a sample is of a certain calendar age or younger than it. It takes the entire calibrated distribution and finds the proportion of the distribution to the right of year x.
}
\details{
The function can only deal with one date at a time
}
\examples{
younger(2800, 2450, 20)
younger(2400, 2450, 20)
calibrate(160, 20, BCAD=TRUE)
younger(1750, 160, 20, BCAD=TRUE)
# we can also use this to calculate older-than probabilities:
older <- 1 - younger(2400, 2450, 20)
older
}
\author{
Maarten Blaauw
}
